snazzy new background, side bounds checks for cowboy, and mouse areas adjusted

+++ Tower Defense/src/Screen.java
@@ -32,7 +32,8 @@ public class Screen extends JPanel implements Runnable
 	public static Player player;
 	public static int prevCommand;
 	public static int command = -1;
-
+	
+	Image bg = new ImageIcon("Resources/gameborder.png").getImage();
 	
 	public Screen(Frame frame)   // for the screen
 	{
@@ -84,7 +85,7 @@ public class Screen extends JPanel implements Runnable
 			if(room.block[r1][c1].groundID == Value.western)
 			{
 				// check for friendly fire!
-				if(room.block[r2][c2-1].groundID >= 30 &&
+				if(c2 - 1 >= 0 && room.block[r2][c2-1].groundID >= 30 &&
 						room.block[r2][c2-1].groundID <= 39)
 				{
 					kill(r2, c2 - 1);
@@ -94,7 +95,7 @@ public class Screen extends JPanel implements Runnable
 				{
 					kill(r2, c2);
 				}
-				if(room.block[r2][c2+1].groundID >= 30 &&
+				if(c2 + 1 < room.block[r2].length && room.block[r2][c2+1].groundID >= 30 &&
 						room.block[r2][c2+1].groundID <= 39)
 				{
 					kill(r2, c2 + 1);
@@ -224,8 +225,8 @@ public class Screen extends JPanel implements Runnable
 			isFirst = false;        //set to not first anymore
 		}
 		g.setColor(new Color(204, 204, 204));
-		g.fillRect(0, 0, getWidth(), getHeight());//clears the rectangle
-		g.setColor(new Color (50,50,50));
+		g.drawImage(bg, 0 , 0, 1220, 550, null);
+		g.setColor(new Color (50, 50, 50));
 		g.drawLine(room.block[0][0].x -1 ,0, room.block[0][0].x-1, room.block[room.worldHeight -1][0].y + room.blockSize); // draw right line
 		g.drawLine(room.block[0][room.worldWidth-1].x + room.blockSize ,0, room.block[0][room.worldWidth-1].x + room.blockSize, room.block[room.worldHeight -1][0].y + room.blockSize);//draw left line
 		g.drawLine(room.block[0][0].x, room.block[room.worldHeight-1][0].y + room.blockSize, room.block[0][room.worldWidth-1].x + room.blockSize,room.block[room.worldHeight-1][0].y +room.blockSize);//draw bottom line
@@ -283,7 +284,7 @@ public class Screen extends JPanel implements Runnable
 					}
 				}
 				addUnit(7, 4, Value.toaster0);
-				addUnit(7, 9, Value.western);
+				addUnit(7, 13, Value.western);
 			}
 			repaint();          // repaints
 			try{
